refactor(ff-decode): split decoder_inner.rs into focused submodules
